+++ lib/x86simdsort-scalar.h
@@ -5,12 +5,12 @@
 namespace xss {
 namespace scalar {
     template <typename T>
-    void qsort(T *arr, int64_t arrsize)
+    void qsort(T *arr, size_t arrsize)
     {
         std::sort(arr, arr + arrsize, compare<T, std::less<T>>());
     }
     template <typename T>
-    void qselect(T *arr, int64_t k, int64_t arrsize, bool hasnan)
+    void qselect(T *arr, size_t k, size_t arrsize, bool hasnan)
     {
         if (hasnan) {
             std::nth_element(arr, arr + k, arr + arrsize, compare<T, std::less<T>>());
@@ -20,7 +20,7 @@ namespace scalar {
         }
     }
     template <typename T>
-    void partial_qsort(T *arr, int64_t k, int64_t arrsize, bool hasnan)
+    void partial_qsort(T *arr, size_t k, size_t arrsize, bool hasnan)
     {
         if (hasnan) {
             std::partial_sort(arr, arr + k, arr + arrsize, compare<T, std::less<T>>());
@@ -30,17 +30,17 @@ namespace scalar {
         }
     }
     template <typename T>
-    std::vector<int64_t> argsort(T *arr, int64_t arrsize)
+    std::vector<size_t> argsort(T *arr, size_t arrsize)
     {
-        std::vector<int64_t> arg(arrsize);
+        std::vector<size_t> arg(arrsize);
         std::iota(arg.begin(), arg.end(), 0);
         std::sort(arg.begin(), arg.end(), compare_arg<T, std::less<T>>(arr));
         return arg;
     }
     template <typename T>
-    std::vector<int64_t> argselect(T *arr, int64_t k, int64_t arrsize)
+    std::vector<size_t> argselect(T *arr, size_t k, size_t arrsize)
     {
-        std::vector<int64_t> arg(arrsize);
+        std::vector<size_t> arg(arrsize);
         std::iota(arg.begin(), arg.end(), 0);
         std::nth_element(arg.begin(),
                          arg.begin() + k,
