+++ lib/x86simdsort-skx.cpp
@@ -6,27 +6,27 @@
 
 #define DEFINE_ALL_METHODS(type) \
     template <> \
-    void qsort(type *arr, int64_t arrsize) \
+    void qsort(type *arr, size_t arrsize) \
     { \
         avx512_qsort(arr, arrsize); \
     } \
     template <> \
-    void qselect(type *arr, int64_t k, int64_t arrsize, bool hasnan) \
+    void qselect(type *arr, size_t k, size_t arrsize, bool hasnan) \
     { \
         avx512_qselect(arr, k, arrsize, hasnan); \
     } \
     template <> \
-    void partial_qsort(type *arr, int64_t k, int64_t arrsize, bool hasnan) \
+    void partial_qsort(type *arr, size_t k, size_t arrsize, bool hasnan) \
     { \
         avx512_partial_qsort(arr, k, arrsize, hasnan); \
     } \
     template <> \
-    std::vector<int64_t> argsort(type *arr, int64_t arrsize) \
+    std::vector<size_t> argsort(type *arr, size_t arrsize) \
     { \
         return avx512_argsort(arr, arrsize); \
     } \
     template <> \
-    std::vector<int64_t> argselect(type *arr, int64_t k, int64_t arrsize) \
+    std::vector<size_t> argselect(type *arr, size_t k, size_t arrsize) \
     { \
         return avx512_argselect(arr, k, arrsize); \
     }
