+++ lib/x86simdsort-spr.cpp
@@ -5,17 +5,17 @@
 namespace xss {
 namespace avx512 {
     template <>
-    void qsort(_Float16 *arr, int64_t size)
+    void qsort(_Float16 *arr, size_t size)
     {
         avx512_qsort(arr, size);
     }
     template <>
-    void qselect(_Float16 *arr, int64_t k, int64_t arrsize, bool hasnan)
+    void qselect(_Float16 *arr, size_t k, size_t arrsize, bool hasnan)
     {
         avx512_qselect(arr, k, arrsize, hasnan);
     }
     template <>
-    void partial_qsort(_Float16 *arr, int64_t k, int64_t arrsize, bool hasnan)
+    void partial_qsort(_Float16 *arr, size_t k, size_t arrsize, bool hasnan)
     {
         avx512_partial_qsort(arr, k, arrsize, hasnan);
     }
